remove PrifileEditPage
