+++ src/components/Form/hooks/types.ts
@@ -1,0 +1,5 @@
+export type FormInputHookObject = {
+  value: string;
+  errorMessage: string | null;
+  isValid: boolean;
+};
